frame: removed unnessecary int cast

+++ src/game/Frame.java
@@ -74,7 +74,7 @@ class Frame extends JFrame {
     private void drawCubes(Graphics g, Lvl lvl) {
         g.setColor(Color.BLUE);
         for (Cube cube : lvl.getCubes()) {
-            g.fillRect((int) cube.getX(), (int) cube.getY(), (int) cube.getBounding().getWidth(), (int) cube.getBounding().getHeight());
+            g.fillRect(cube.getX(), cube.getY(), (int) cube.getBounding().getWidth(), (int) cube.getBounding().getHeight());
         }
     }
